+++ RELEASE_NOTES.md
@@ -28,3 +28,4 @@ _Note: Remember to restart the AppDaemon addon/server after updating to a new ve
 ## :video_game: New devices
 
 - [E2002](https://BASE_URL/controllerx/controllers/E2002) - add Media Player support for ZHA.
+- [SK5700002228949](https://BASE_URL/controllerx/controllers/SK5700002228949) - add Light support for deCONZ.